Moves global declaration to setup-env
